fix: resolve react and react-dom dependency conflicts
The specification of react and react-dom @18 causes npm to pull a
version of react at that version within the package directory. This
causes a runtime issue if this is not the version of react that is
actually used by consumers when they attempt to serve those dependencies
without the version matching. `peerDependencies` suggests that this package
should be compatible with react and react-dom @17 as well, so the
expectation is that consumers can use v17 of those dependencies.

`peerDependencies` already specify react and react-dom @17/@18,
so the necessity of these packages is already conveyed to consuming
packages. `devDependencies` already specifies react and react-dom @18
for use for development of this package.

Closes #26